feat: enhance database schema and operations for doctors, including creation, deletion, and updates

+++ src/db/dbDelete.ts
@@ -3,6 +3,7 @@ import { db } from "@/lib/db";
 export const deletePatient = async (id: string) => {
   try {
     if (!id) return;
+
     await db.transact(db.tx.patients[id].delete());
   } catch (error) {
     console.error(error);
@@ -10,9 +11,21 @@ export const deletePatient = async (id: string) => {
 };
 
 export const deletePatientHistoryItem = (id: string) => {
-  db.transact(db.tx.healthCard[id].delete());
+  try {
+    if (!id) return;
+
+    db.transact(db.tx.healthCard[id].delete());
+  } catch (error) {
+    console.error(error);
+  }
 };
 
 export const deleteDoctor = async (id: string) => {
-  db.transact(db.tx.doctors[id].delete());
+  try {
+    if (!id) return;
+
+    db.transact(db.tx.doctors[id].delete());
+  } catch (error) {
+    console.error(error);
+  }
 };
